requirements.txt
psycopg2-binary==2.9.11
feedparser==6.0.12
requests==2.32.5
python-dotenv==1.2.2
beautifulsoup4==4.14.3
python-dateutil==2.9.0.post0
google-generativeai==0.8.3
